Critical HTS functions marked external (#127)
Changed functions perform critical operations that change token state. If they are left external, it would allow anyone to cause very serious issues if the inheriting contract has the ADMIN, KYC, or PAUSE keys for Hedera token[s].

In your tutorials, devs are suggested to inherit these HTS contracts. If they decide to make the inheriting contract have the admin key to allow it to change the token's certain properties behind restricted functions, anyone can bypass the restrictions via the inherited HTS contract, and hijack the token. Or if they make the inheriting contract have the pause key and make it algorithmically manage the pause state of the token, anyone can bypass such algorithm and DOS the token. Given the expected usage of these contracts, and the severity of the impact, this is a high risk issue.

Signed-off-by: georgi-l95 <glazarov95@gmail.com>